requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
pyactiveresource==2.2.1
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
ShopifyAPI==8.0.3
six==1.15.0
urllib3==1.25.10
